+++ src/Main.java
@@ -5,13 +5,13 @@ public class Main {
         double percentYear = 9.99;
         int termDate;
         int monthAmount;
-        int CreditPayment = (int) service.calculate(1_000_000, 12, 1);
+        double CreditPayment = service.calculate(1_000_000, 12, 1);
         System.out.println("Ежемесячный платеж по кредиту на сумму 1000000 сроком на один год составляет: " + CreditPayment);
 
-        int CreditPayment2 = (int) service.calculate(1_000_000, 24, 2);
+        double CreditPayment2 = service.calculate(1_000_000, 24, 2);
         System.out.println("Ежемесячный платеж по кредиту на сумму 1000000 сроком на два года составляет: " + CreditPayment2);
 
-        int CreditPayment3 = (int) service.calculate(1_000_000, 36, 3);
+        double CreditPayment3 = service.calculate(1_000_000, 36, 3);
         System.out.println("Ежемесячный платеж по кредиту на сумму 1000000 сроком на три года составляет: " + CreditPayment3);
     }
 }
